Add eight _Opponent liberties after move_ features
This seems to result in about a 0.5% improve in the policy
accuracy, and about a 0.35% value improvement. It also seems to
result in a 1.0% improve in the policy top 5, which has less of
an obvious effect.

In a quick comparisons of actual play strength, the 40 features
version wins 5 - 1 over the previous 32 features networks.