Initial version of Trainer V2 and the Graph Inference framework
Major upgrade of both the TensorFlow training script, and the graph
inference framework to adopt new technologies:

# Training Script

- Upgrade to TensorFlow 2.0.0a0
- Use the new Keras API instead of Estimators
- Update the model definition with new concepts:
  - Squeeze Excite
  - Global Average Pooling (instead of dense layers before
    classification)
  - Some more additional training targets

The current version of TensorFlow is **very** slow, the nightly build
is much faster but seems to have a memory leak at the moment.

# Inference Engine

Complete re-write of the inference engine in Rust in order to
separate (i) unsafe CUDA code, (ii) layer definitions, and (iii) the
model architecture. The result is a fairly flexible graph execution
engine that reads the model definition from a JSON file and executes
it (like Theano, but without a backward pass).

A few other nice improvements that snuck in this commit:

- Get ride of the cuBLAS dependency (by using cuDNN for the matrix
  multiplication).
- Use CUDA Graphs to speed up repetitive tasks during inference.