Split data into a training and evaluation set during training
When running the training loop, split the data into two sets. One
training set which is used to train the neural network, and an
evaluation set which is used to verify the performance of the neural
network.

This allows us to get an unbiased opinion of when the network starts
overfitting.