Add the `--reanalyze` command
Implement the _Reanalyze_ proceedure described in MuZero which takes an
SGF files, and reanalyze x% of the positions and output new policies and
value estimates that can be learned against.

Currently we only learn based on the policy emitted from the reanalyze
and not the value itself in order to ensure we maintain a fixed ground
truth.