Semi-supervised learning v1.0
Introduce a semi-supervised mode that can be used with leela-zero
weights. When used the an additional loss is added to the optimization
that minimize the difference between the final internal representation
of the neural networks.

This is preferred to setting the labels directly from the leela-zero
network because that does not ensure a well formed internal
representation.