New library for streaming JSON parsing
Trivial speed-up in the JSON weights parsing by re-writing the loader
using a less-copy (almost zero-copy, but not quite) parser.

Most of the weights parsing time is still spent in base-85 decoding so
this does not make much of a difference, but makes future JSON files
more viable since we have a general parser.