Implement weight averaging during training
When dumping the weights, dump the average weights of the last 4,000
iterations (using a moving average algorithm with Adam-style zero
debias). We do not need to bother with the _batch norm fix_ since we
calculate the moving average on the folded weights :-)

This implementation differs from LZ / KataGo SWA, but the results should
be very similar.

The engine seems to be slightly stronger, when playing with one second
per move time control, when using weight averaging than without, but
also not a massive difference:

```
9x128-5522cac v 9x128-5522cac-moving_avg (200/200 games)
board size: 19   komi: 7.5
                           wins              black         white        avg cpu
9x128-5522cac                95 47.50%       45 45.00%     50  50.00%   2116.33
9x128-5522cac-moving_avg    105 52.50%       50 50.00%     55  55.00%   2070.87
                                             95 47.50%     105 52.50%
```