Remove incorrect quantization during training
During training we used to quantize the `initial_state` (`h`) and the
`embeddings` (`x`). This is incorrect, we want to quantize the
`initial_state` (`h`) and any intermediate hidden states. Until we know
how to do this (probably will require a custom `GRUCell` implementation)
remove the quantization.

This fixes the `model` unit tests.