Implement support for the dynamics network in `dg_mcts`
Call the dynamics network during monte-carlo tree search instead of the
representation network every time. This should be a bit faster since the
dynamics network should in-theory be a bit smaller.

The implementation is a bit ugly, since it is the fastest hack I could
think of to test it. But it should work and be _good enough_ until we
rewrite the monte-carlo tree search crate.